Move Logo in atoms
