+++ src/pages/Login/Login.jsx
@@ -9,7 +9,7 @@ import Grid from '@mui/material/Grid';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
-import { Logo } from '../../components/molecules/Logo/Logo';
+import { Logo } from '../../components/atoms/Logo/Logo';
 
 function Copyright(props) {
   return (
